+++ src/Event/SocialShare.php
@@ -6,6 +6,19 @@ use Pagekit\Event\EventSubscriberInterface;
 
 class SocialShare implements EventSubscriberInterface{
 
+     /**
+     * @var Module
+     */
+    protected $blog;
+
+    /**
+     * Constructor.
+     */
+    public function __construct()
+    {
+        $this->blog = App::module('dpnblog');
+    }
+
     public function onContent(ContentEvent $event){
         if (!$event['post']) {
             return;
@@ -20,7 +33,10 @@ class SocialShare implements EventSubscriberInterface{
             'desc' => isset($post->data['meta']['og:description']) ? $post->data['meta']['og:description']:$post->excerpt,
         ];
         $content = $event->getContent();
-        $reply = App::view('dpnblog/components/social.php', compact(['data']));
+        $reply ='';
+        if ( $this->blog->config('posts.social_share_enabled')){
+            $reply = App::view('dpnblog/components/social.php', compact(['data']));
+        }
         $event->setContent($content . $reply);
     }