+++ views/components/social.php
@@ -8,8 +8,5 @@
         <div>
             <a href="https://twitter.com/home?status=<?= $app['url']->current(0) ?>" target="_blank" class="uk-button uk-button-default"><span class="uk-icon uk-icon-twitter tm-icon-twitter"></span> Twitter Share</a>
         </div>
-        <div>
-            <a href="https://plus.google.com/share?url=<?= $app['url']->current(0) ?>" target="_blank" class="uk-button uk-button-default"><span class="uk-icon uk-icon-google-plus tm-icon-google"></span> Google + Share</a>
-        </div>
     </div>
 </section>