+++ views/post.php
@@ -34,9 +34,13 @@
                 <?php endforeach; ?>
             </ul>
         </div>
+
+        <!-- <div class="uk-margin-small">< ?= $post->excerpt ?></div> -->
+
         <?php if ($authorBox === true): ?>
             <div class="uk-comment uk-margin-bottom">
                 <div class="uk-comment-header">
+                    <!-- <img class="dpnblog-avatar-comment" src="< ?= $post->getGravatar() ?>"> -->
                     <img class="uk-comment-avatar" src="<?= $post->getGravatar() ?>">
                     <h4 class="uk-comment-title"><?= $user->name ?></h4>
                     <div class="uk-comment-meta">