+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 Any changes to this project will be written to this document. With a detailed description, you can easily see what changes in V2.
 
+
+|*November 08, 2019 - 13:30PM*
+
+### *Added*
+
+    - Post Widget (display list of posts)
+    - Setting for Social Share buttons (Enable or not)
+    - litl redisiened
+    - beta ru_RU language
+
 *April 08, 2019 - 17:30PM*
 
 ### *Fixed*